Updated inline Documentation for this Class
I have added more detailed comments to this class file as well as updated the type hinting on each function to help prevent headaches in the future.

Some Areas type hinting was not added as it would require more time to adjust to work correctly with the old checking code in in the functions.